+++ cognite/neat/graph/transformers/__init__.py
@@ -1,5 +1,5 @@
-from ._classic_cdf import AddAssetDepth
+from ._classic_cdf import AddAssetDepth, AssetTimeSeriesConnector
 
-__all__ = ["AddAssetDepth"]
+__all__ = ["AddAssetDepth", "AssetTimeSeriesConnector"]
 
-Transformers = AddAssetDepth
+Transformers = AddAssetDepth | AssetTimeSeriesConnector